requirements.txt
h5py==3.1.0
imageio==2.9.0
matplotlib==3.3.3
numpy==1.22.1
pandas==1.2.3
Pillow==9.1.0
PyYAML==6.0
torch==1.7.1+cu110
torchvision==0.8.2+cu110
transformers==4.17.0
